Attempt to emit deprecation warning only when _agg param is used/modified
General scripted metric agg use requires modifying _agg to track state per document. My rationale is
that most (all?) practical uses of the feature require _agg to be modified in the init or map scripts.

The benefit of this is that it will not emit the deprecation warning if scripted metric aggs are used
with the new state context variable. The downside is that it will not emit a deprecation warning in the
case where _agg is never modified in map/init/combine (if that is in fact a practical use case).